+++ src/button/button.ts
@@ -92,6 +92,9 @@ export default class Button extends SuperComponent {
     launchapp(e) {
       this.triggerEvent('launchapp', e.detail);
     },
+    chooseavatar(e) {
+      this.triggerEvent('chooseavatar', e.detail);
+    },
     handleTap(e) {
       if (this.data.disabled) return;