+++ src/loading/type.ts
@@ -2,7 +2,6 @@
 
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * updated at 2021-11-24 10:58:05
  * */
 
 export interface TdLoadingProps {
@@ -13,7 +12,6 @@ export interface TdLoadingProps {
   delay?: {
     type: NumberConstructor;
     value?: number;
-    required?: boolean;
   };
   /**
    * 加载动画执行完成一次的时间，单位：毫秒
@@ -22,7 +20,6 @@ export interface TdLoadingProps {
   duration?: {
     type: NumberConstructor;
     value?: number;
-    required?: boolean;
   };
   /**
    * 组件类名，分别用于设置加载组件外层元素，加载组件文本，加载组件指示符，加载指示符内侧同心圆等元素类名
@@ -30,7 +27,6 @@ export interface TdLoadingProps {
   externalClasses?: {
     type: ArrayConstructor;
     value?: ['t-class', 't-class-text', 't-class-indicator'];
-    required?: boolean;
   };
   /**
    * 是否显示加载指示符
@@ -39,7 +35,14 @@ export interface TdLoadingProps {
   indicator?: {
     type: BooleanConstructor;
     value?: boolean;
-    required?: boolean;
+  };
+  /**
+   * 是否继承父元素颜色
+   * @default false
+   */
+  inheritColor?: {
+    type: BooleanConstructor;
+    value?: boolean;
   };
   /**
    * 对齐方式
@@ -48,7 +51,6 @@ export interface TdLoadingProps {
   layout?: {
     type: StringConstructor;
     value?: 'horizontal' | 'vertical';
-    required?: boolean;
   };
   /**
    * 是否处于加载状态
@@ -57,7 +59,6 @@ export interface TdLoadingProps {
   loading?: {
     type: BooleanConstructor;
     value?: boolean;
-    required?: boolean;
   };
   /**
    * 是否暂停动画
@@ -66,7 +67,6 @@ export interface TdLoadingProps {
   pause?: {
     type: BooleanConstructor;
     value?: boolean;
-    required?: boolean;
   };
   /**
    * 加载进度
@@ -74,7 +74,6 @@ export interface TdLoadingProps {
   progress?: {
     type: NumberConstructor;
     value?: number;
-    required?: boolean;
   };
   /**
    * 加载动画是否反向
@@ -82,7 +81,6 @@ export interface TdLoadingProps {
   reverse?: {
     type: BooleanConstructor;
     value?: boolean;
-    required?: boolean;
   };
   /**
    * 尺寸，示例：40rpx/20px
@@ -91,7 +89,6 @@ export interface TdLoadingProps {
   size?: {
     type: StringConstructor;
     value?: string;
-    required?: boolean;
   };
   /**
    * 加载提示文案
@@ -99,7 +96,6 @@ export interface TdLoadingProps {
   text?: {
     type: StringConstructor;
     value?: string;
-    required?: boolean;
   };
   /**
    * 加载组件类型
@@ -108,6 +104,5 @@ export interface TdLoadingProps {
   theme?: {
     type: StringConstructor;
     value?: 'circular' | 'spinner' | 'bar' | 'error' | 'dots';
-    required?: boolean;
   };
-};
+}