+++ src/toast/props.ts
@@ -2,7 +2,6 @@
 
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * updated at 2021-11-24 10:58:05
  * */
 
 import { TdToastProps } from './type';
@@ -30,6 +29,11 @@ const props: TdToastProps = {
   message: {
     type: String,
   },
+  /** 遮罩层属性，透传至 Overlay */
+  overlayProps: {
+    type: Object,
+    value: {},
+  },
   /** 弹窗展示位置 */
   placement: {
     type: String,
@@ -37,6 +41,11 @@ const props: TdToastProps = {
   },
   /** 防止滚动穿透，即不允许点击和滚动 */
   preventScrollThrough: {
+    type: Boolean,
+    value: false,
+  },
+  /** 是否显示遮罩层 */
+  showOverlay: {
     type: Boolean,
     value: false,
   },